+++ lizard_rainapp/admin.py
@@ -1,0 +1,4 @@
+from django.contrib import admin
+from lizard_rainapp.models import Setting
+
+admin.site.register(Setting)